Switched Surface functions to use float precision.

+++ ENIGMAsystem/SHELL/Graphics_Systems/DirectX10/DX10blend.cpp
@@ -25,7 +25,7 @@ int draw_set_blend_mode(int mode){
 
 }
 
-int draw_set_blend_mode_ext(double src,double dest){
+int draw_set_blend_mode_ext(int src, int dest){
 
 }
